Update runprogram_endinstance_withnotice.sh
NOTE: this requires that the AWS account this script is being used in has set up an origination number. To do this go to https://console.aws.amazon.com/pinpoint/home?region=us-east-1#/sms-account-settings/requestLongCode and set up a toll-free number that the SMS message can be sent from. 

This script runs the block of code you would like to run (put in between the lines under step 4), sends an SMS message to the phone number entered, and then stops or terminates the EC2 instance.